tests/requirements.txt
# Requirements for test scripts
psycopg2-binary>=2.9.5
aiohttp>=3.8.3
pytest>=7.0.0
pytest-asyncio>=0.18.0
fastapi>=0.100.0
requests>=2.31.0
httpx>=0.24.0
SQLAlchemy>=2.0.0
websockets>=11.0.0
pytest-cov>=4.1.0